@@ -1,6 +1,6 @@
 import { Recipe } from './../recipes/recipe.model';
 import { RecipeService } from './../recipes/recipe.service';
-import { Http, Response } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import 'rxjs/Rx';
 
@@ -8,7 +8,7 @@ import 'rxjs/Rx';
 export class DataStorageService {
   API_URL = 'https://ng-recipe-book-c67db.firebaseio.com/recipes.json';
 
-  constructor(private http: Http, private recipeService: RecipeService) {}
+  constructor(private http: HttpClient, private recipeService: RecipeService) {}
 
   storeRecipes() {
     return this.http.put(this.API_URL, this.recipeService.getRecipes());
@@ -17,8 +17,7 @@ export class DataStorageService {
   getRecipes() {
     this.http.get(this.API_URL)
     .map(
-      (response: Response) => {
-        const recipes: Recipe[] = response.json();
+      (recipes: Recipe[]) => {
         for (const recipe of recipes){
           if (!recipe['ingredients']) {
             recipe['ingredients'] = [];
